% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-qubic.R
\name{BCQU-class}
\alias{BCQU-class}
\title{Class BCQU.}
\value{
Class generator for objects of class \code{BCQU}.
}
\description{
Class \code{BCQU} define a QUalitative BIClustering calcuator.
}
\seealso{
\code{\link{BCQU}} \code{\link{qudiscretize}} \code{\link{qunetwork}} \code{\link{qunet2xml}}
}
