% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qunet2xml.R
\name{qunet2xml}
\alias{qunet2xml}
\title{Convert newwork to XGMML}
\usage{
qunet2xml(
  net,
  minimum = 0.6,
  color = cbind(grDevices::rainbow(length(net[[2]]) - 1), "gray")
)
}
\arguments{
\item{net}{Result of \code{\link{qunetwork}}}

\item{minimum}{cutoff, default: 0.6}

\item{color}{default: cbind(grDevices::rainbow(length(net[[2]]) - 1), 'gray')}
}
\value{
Text of XGMML
}
\description{
This function can convert the constructed co-expression networks into XGMML format, which can be used to do further network analysis in Cytoscape, Biomax and JNets.
}
\examples{
# Load microarray matrix
if (requireNamespace("QUBICdata", quietly = TRUE)) {
  data(ecoli, package = "QUBICdata")
  res <- qubiclust(ecoli, r = 1, q = 0.06, c = 0.95,
                  o = 100, f = 0.25, k = max(ncol(ecoli) \%/\% 20, 2),
                  verbose = FALSE)
  # Get all biclusters
  net <- qunetwork(ecoli, res, number = 5, groups = 5, method = "spearman")
  # Save the network to a XGMML file
  outfile <- tempfile(fileext = ".gr")
  sink(outfile)
  qunet2xml(net, minimum = 0.6,
           color = cbind(grDevices::rainbow(length(net[[2]]) - 1), "gray"))
  sink()
  unlink(outfile)
}
}
\seealso{
\code{\link{qunetwork}} \code{\link{QUBIC}}
}
