% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qunetwork.R
\name{qunetwork}
\alias{qunetwork}
\alias{Qnetwork}
\alias{network}
\alias{Qunetwork}
\title{Construction and visualization of co-expression network}
\usage{
qunetwork(
  x,
  BicRes,
  number = NULL,
  groups = NULL,
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x}{The data matrix}

\item{BicRes}{Result object returned by \code{\link{qubiclust}} or
\code{\link{qubiclust_d}}}

\item{number}{Which bicluster to be plotted}

\item{groups}{An object that indicates which nodes belong together.}

\item{method}{A character string indicating
which correlation coefficient (or covariance) is to be computed.
One of 'pearson' (default), 'kendall', or 'spearman', can be abbreviated.}
}
\value{
a list contains a weights matrix and groupinfo
}
\description{
This function can automatically create co-expression networks along with their visualization based on identified biclusters in QUBIC.
Three correlation methods, Pearson, Kendall and Spearman, are available for a user, facilitating different preferences in practical usage.
}
\examples{
# Load microarray matrix
if (requireNamespace("QUBICdata", quietly = TRUE)) {
  data(ecoli, package = "QUBICdata")
  res <- qubiclust(ecoli, r = 1, q = 0.06, c = 0.95,
                  o = 100, f = 0.25, k = max(ncol(ecoli) \%/\% 20, 2),
                  verbose = FALSE)
  # Constructing the network for the 5th identified bicluster.
  net <- qunetwork(ecoli, res, number = 5, groups = 5, method = "spearman")
}
\dontrun{
if (requireNamespace('qgraph'))
    qgraph::qgraph(net[[1]], groups = net[[2]], layout = 'spring', minimum = 0.6,
       color = cbind(rainbow(length(net[[2]]) - 1),'gray'), edge.labels = FALSE)

}
\dontrun{
# Load microarray matrix
data(ecoli, package = "QUBICdata")
res <- qubiclust(ecoli, r = 1, q = 0.06, c = 0.95,
                o = 100, f = 0.25, k = max(ncol(ecoli) \%/\% 20, 2),
                verbose = FALSE)
# Constructing the networks for the 4th and 8th identified biclusters,
#   using the whole network as a background.
net <- qunetwork(ecoli, res, number = c(4, 8), groups = c(4, 8), method = "spearman")
if (requireNamespace('qgraph'))
    qgraph::qgraph(net[[1]], groups = net[[2]], layout = 'spring', minimum = 0.6,
       color = c('red', 'blue', 'gold', 'gray'), edge.labels = FALSE)
}
}
\seealso{
\code{\link{qunet2xml}} \code{\link{QUBIC}} \code{\link{cor}}
}
