% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showinfo.R
\name{showinfo}
\alias{showinfo}
\title{Show report of biclusters}
\usage{
showinfo(matrix, bic)
}
\arguments{
\item{matrix}{microarray matrix}

\item{bic}{array of biclusters}
}
\value{
Text of report
}
\description{
This function can make a report for biclusters.
}
\examples{
# Load microarray matrix
if (requireNamespace("QUBICdata", quietly = TRUE)) {
  data(ecoli, package = "QUBICdata")
  matrix <- ecoli[1:50, , drop = FALSE]
  res1 <- qubiclust(matrix, verbose = FALSE)
  res2 <- qubiclust(matrix, c = 0.9, verbose = FALSE)
  res3 <- qubiclust(matrix, c = 0.85, verbose = FALSE)
  # Show the report
  showinfo(matrix, list(res1, res2, res3))
}
}
\seealso{
\code{\link{QUBIC}}
}
