% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-qubic.R
\docType{class}
\name{BCQU-class}
\alias{BCQU-class}
\title{Class BCQU.}
\description{
Class \code{BCQU} define a QUalitative BIClustering calcuator.
}
\seealso{
\code{\link{BCQU}} \code{\link{qudiscretize}} \code{\link{qunetwork}} \code{\link{qunet2xml}} \code{\link{biclust}}
}

