% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showinfo.R
\name{showinfo}
\alias{showinfo}
\title{Show report of biclusters}
\usage{
showinfo(matrix, bic)
}
\arguments{
\item{matrix}{microarray matrix}

\item{bic}{array of biclusters}
}
\value{
Text of report
}
\description{
This function can make a report for biclusters.
}
\examples{
# Load microarray matrix
data(BicatYeast)
matrix <- BicatYeast[1:50, ];
res1 <- biclust::biclust(matrix, method=BCQU(), verbose = FALSE)
res2 <- biclust::biclust(matrix, method=BCCC())
res3 <- biclust::biclust(matrix, method=BCBimax())
# Show the report
showinfo(matrix, c(res1, res2, res3))
}
\seealso{
\code{\link{QUBIC}}
}

